fix(pickers): test plan picker shows "unhandled resource" lines
This fixes an issue similar to #326 that happens when selecting a test
plan and one of the embedded SPM packages prints those warnings, making
them appear in the list of test plans.

Unfortunately these lines are just printed to stdout by spm and I
couldn't find a more direct way of filtering them out.